+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-import roomescape.dao.resetvationTime.ReservationTimeDao;
+import roomescape.dao.reservationTime.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
 import roomescape.dto.request.ReservationTimeCreateRequest;
 import roomescape.dto.response.ReservationTimeCreateResponse;
