+++ src/test/java/roomescape/dao/reservationTime/JdbcReservationTimeDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao.resetvationTime;
+package roomescape.dao.reservationTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
